enum package renamed to type
