+++ android/TraderX/app/src/main/java/com/traderx/type/AlertType.kt
@@ -1,4 +1,4 @@
-package com.traderx.enum
+package com.traderx.type
 
 enum class AlertType(val request: String, val response: String) {
     ABOVE("above", "ABOVE"),
