+++ android/TraderX/app/src/main/java/com/traderx/type/TransactionType.kt
@@ -1,4 +1,4 @@
-package com.traderx.enum
+package com.traderx.type
 
 enum class TransactionType(val request: String, val response: String) {
     BUY("buy", "BUY"),
