+++ android/TraderX/app/src/main/java/com/traderx/ui/equipment/EquipmentsFragment.kt
@@ -12,7 +12,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.traderx.R
 import com.traderx.api.ErrorHandler
 import com.traderx.api.response.EquipmentsResponse
-import com.traderx.enum.EquipmentType
+import com.traderx.type.EquipmentType
 import com.traderx.ui.search.UserSearchSkeletonRecyclerViewAdapter
 import com.traderx.util.FragmentTitleEmitters
 import com.traderx.util.FragmentTitleListeners
